# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/feeds": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can see feeds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Post"
                ],
                "summary": "User can see feeds",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "searchkey",
                        "name": "searchkey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.UserFeedsResponse"
                        }
                    }
                }
            }
        },
...
        "/user/post/details/report": {
            "post": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can Report post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Post"
                ],
                "summary": "User can report a post",
                "parameters": [
                    {
                        "description": "Report Post Data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.ReportPostRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.ReportPostResponse"
                        }
                    }
                }
            }
        },
        "/user/post/donate/": {
            "post": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can donate for campaigns",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Donations"
                ],
                "summary": "User can donate for campaign",
                "parameters": [
                    {
                        "description": "Donate Data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.DonateRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.DonateResponse"
                        }
                    }
                }
            }
        },
...
        "/user/post/donate/history": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can see Donation History",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Donations"
                ],
                "summary": "User can see Donation History",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "searchkey",
                        "name": "searchkey",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.DonationHistoryResponse"
                        }
                    }
                }
            }
        },
...
        "/user/post/donate/razorpay": {
            "get": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can pay for donation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Donations"
                ],
                "summary": "User can pay for donation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pay id Data",
                        "name": "payid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.MakePaymentRazorPayResponse"
                        }
                    }
                }
            }
        },
...
        "/user/post/like": {
            "post": {
                "security": [
                    {
                        "api_key": []
                    }
                ],
                "description": "User can Like post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Post"
                ],
                "summary": "User can Like a post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PostID",
                        "name": "postid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.LikePostResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.Comment": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "postid": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                },
                "userid": {
                    "type": "integer"
                }
            }
        },
...
        "pb.DonateResponse": {
            "type": "object",
            "properties": {
                "Post": {
                    "$ref": "#/definitions/pkg_user_pb.Post"
                },
                "link": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "pb.Donation": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paymentid": {
                    "type": "string"
                },
                "post": {
                    "$ref": "#/definitions/pkg_user_pb.Post"
                }
            }
        },
        "pb.DonationHistoryResponse": {
            "type": "object",
            "properties": {
                "donations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Donation"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.LikePostResponse": {
            "type": "object",
            "properties": {
                "Post": {
                    "$ref": "#/definitions/pb.PostDetails"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.MakePaymentRazorPayResponse": {
            "type": "object",
            "properties": {
                "FinalPrice": {
                    "type": "integer"
                },
                "PaymentID": {
                    "type": "integer"
                },
                "Razor_id": {
                    "type": "string"
                },
                "UserID": {
                    "type": "integer"
                },
                "Username": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.PostDetails": {
            "type": "object",
            "properties": {
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Comment"
                    }
                },
                "post": {
                    "$ref": "#/definitions/pkg_user_pb.Post"
                }
            }
        },
...
        "pb.ReportPostResponse": {
            "type": "object",
            "properties": {
                "Post": {
                    "$ref": "#/definitions/pkg_user_pb.Post"
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pb.UserFeedsResponse": {
            "type": "object",
            "properties": {
                "Posts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_user_pb.Post"
                    }
                },
                "response": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "pkg_user_pb.Post": {
            "type": "object",
            "properties": {
                "account_no": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "amount": {
                    "type": "integer"
                },
                "collected": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "likes": {
                    "type": "integer"
                },
                "place": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "routes.DonateRequestBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "postID": {
                    "type": "integer"
                }
            }
        },
...
        "routes.ReportPostRequestBody": {
            "type": "object",
            "properties": {
                "postId": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    }
}